# posuto-server/src/models/PostModel.js
import mongoose from 'mongoose';

// const PositionSchema = new mongoose.Schema({
//   'gs-w': String,
//   'gs-h': String,
//   'gs-x': String,
//   'gs-y': String,
// });
const postSchema = new mongoose.Schema(
  {
    title: {
      type: String,
      required: true,
      trim: true,
      maxlength: 50,
    },
    contents: String,
    position: {
      'gs-w': Number,
      'gs-h': Number,
      'gs-x': Number,
      'gs-y': Number,
    },
    color: String,
    createdBy: {
      type: mongoose.SchemaTypes.ObjectId,
      ref: 'Users',
      required: true,
    },
  },
  { timestamps: true },
);

postSchema.index({ Users: 1, title: 1 }, { unique: true });
const PostModel = mongoose.model('Posts', postSchema);

export default PostModel;